requirements.txt
termcolor~=2.5.0
pyfiglet~=1.0.2
mysql-connector-python~=8.0.33

